Add 'user add' command
Add the 'user add' command.
Check the request method and pass it to the HMAC signer so that it can sign
the request correctly.
Add a debug (-d) flag to the cli and use it to set the client debug level.
Add a debug? method to the Thor class so we can use it to detect when debug is
turned on.