Add 'organizations' sub-command
Add the 'organizations' sub-command with 'org' as an alias.
Copy modified forms of the 'admin org show' & 'admin org modify' into the
'organizations' sub-command that operate only on the current organization
defined in the config file.
Extract the 'organization' field from the config file and provide a reader
attribute in the Config.