Abstract the configuration & API authentication
Use an options Hash for the config class; allow the configuration file path to
be optional and the config to be passed in as Hash keys, too.
Uncouple the API HTTP request methods; rename the signed_* methods to the more
generic api_* methods, and proxy requests to the configured API authentication
type methods.
Move the signed_ methods & related HMAC methods to the Client::Http::HMAC
class.
Move the HTTP related methods to the Client::Http Module.