Improve TLS support
Allow the server URL to be specified as a single url in the config file and
parse out the scheme, server & port from it if it's there.
Add the tls option to the config file.
If tls is enabled ("tls: true" or the scheme is "https") always enable TLS
during API operations.
Make sure "org list" always displays the correct URL including scheme, and
rename "Server" to "URL"