Add the 'job format' command
Add the #format method.
Rationalise the job loading/parsing code into two provide methods so we can
re-use it from the now-three methods that all want to load a job.
Trusty ships with Ruby 2.0.0-p484, which is now EOL and too old to work
reliably. Install Ruby 2.1.0 from the Brightbox PPA instead, which is a
minimum sensible version.